First working version of treeView
