+++ CodeAnalyzerTool/CodeAnalyzerService.UI/code-analyzer-service.ui/src/components/AnalysisSummary.tsx
@@ -1,25 +1,18 @@
 import {
     Alert, Container, Divider,
-    FormControl, Grid, List, ListItem, ListItemButton, ListItemText, MenuItem, Select, SelectChangeEvent,
+    FormControl, Grid, MenuItem, Select, SelectChangeEvent,
     Typography
 } from "@mui/material";
-import TreeView from '@mui/lab/TreeView';
 import {
     createFolderHierarchy,
-    Folder,
     groupRuleViolationsByPath,
-    splitPathsIntoStringArray
-} from "../util/HelperFunctions";
+} from "../services/ruleViolationService";
 import React, {useState} from "react";
 import {useAnalysis} from "../hooks/useAnalysis";
 import Loading from "./Loading";
 import {AnalysisHistory} from "../model/Analysis";
 import FileViolations from "./FileViolations";
 import DonePlaceholder from "./placeholders/DonePlaceholder";
-import ExpandMoreIcon from '@mui/icons-material/ExpandMore';
-import ChevronRightIcon from '@mui/icons-material/ChevronRight';
-import {TreeItem} from "@mui/lab";
-import RenderTreeView from "./RenderTreeView";
 import { FolderTree } from "./FolderTree";
 
 interface AnalysisSummaryProps {
@@ -37,10 +30,9 @@ export default function AnalysisSummary({initialAnalysisId, analysisHistory}: An
     if (isError || !analysis){
         return <Alert severity="error">Error loading the analysis</Alert>
     }
+
     const groupedAnalysis = groupRuleViolationsByPath(analysis)
-    // const folderHierarchy: Folder = createFolderHierarchy(["code-analyzer-tool\\CodeAnalyzerTool\\CodeAnalyzerService.Backend\\BL\\Services\\RuleService.cs", "code-analyzer-tool\\CodeAnalyzerTool\\CodeAnalyzerService.Backend\\Controllers\\RuleViolationController.cs"])
-    const folderHierarchy: Folder = createFolderHierarchy(Object.keys(groupedAnalysis))
-    console.log(folderHierarchy)
+    const folderHierarchy = createFolderHierarchy(Object.keys(groupedAnalysis))
 
 
     const handleChange = (event: SelectChangeEvent) => {
@@ -89,13 +81,7 @@ export default function AnalysisSummary({initialAnalysisId, analysisHistory}: An
                     {Object.entries(groupedAnalysis).map(([path, violations], index) => (
                         <FileViolations id={`path-${index}`} key={index} path={path} violations={violations}/>
                     ))}
-                    {/*<TreeView aria-label="file system navigator"*/}
-                    {/*          defaultCollapseIcon={<ExpandMoreIcon />}*/}
-                    {/*          defaultExpandIcon={<ChevronRightIcon />}*/}
-                    {/*          sx={{ height: 240, flexGrow: 1, maxWidth: 400, overflowY: 'auto' }}>*/}
-                    {/*    {folderHierarchy.children && folderHierarchy.children.map((child) => RenderTreeView(child))}*/}
-                    {/*</TreeView>*/}
-                    <FolderTree folder={folderHierarchy} />
+                    {folderHierarchy && folderHierarchy.length > 0 && <FolderTree folders={folderHierarchy}/>}
                 </>
             }
         </Container>
